Subscription refactoring and docs
